# docs/swagger.yaml
definitions:
...
  models.Cart:
    properties:
      created_at:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.CartItem:
    properties:
      created_at:
        type: string
      product:
        $ref: '#/definitions/models.ProductWeight'
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.ProductWeight:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      price:
        type: number
      productID:
        type: integer
      stockAvailability:
        type: integer
      weight:
        type: integer
    type: object
...
  services.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /cart/{userID}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User's cart retrieved successfully
          schema:
            $ref: '#/definitions/models.Cart'
      summary: Get the user's cart by user ID
      tags:
      - Cart
    post:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Cart item object
        in: body
        name: cartItem
        required: true
        schema:
          $ref: '#/definitions/models.CartItem'
      produces:
      - application/json
      responses:
        "200":
          description: Item added to cart successfully
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/services.ErrorResponse'
      summary: Add an item to the cart
      tags:
      - Cart
  /cart/{userID}/{productWeightID}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Product Weight ID
        in: path
        name: productWeightID
        required: true
        type: string
      responses:
        "200":
          description: Cart item removed successfully
          schema:
            type: string
        "404":
          description: Cart item not found
          schema:
            $ref: '#/definitions/services.ErrorResponse'
      summary: Remove a cart item from the cart
      tags:
      - Cart
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Product Weight ID
        in: path
        name: productWeightID
        required: true
        type: string
      - description: Cart item object
        in: body
        name: cartItem
        required: true
        schema:
          $ref: '#/definitions/models.CartItem'
      produces:
      - application/json
      responses:
        "200":
          description: Cart item updated successfully
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/services.ErrorResponse'
        "404":
          description: Cart item not found
          schema:
            $ref: '#/definitions/services.ErrorResponse'
      summary: Update the quantity of a cart item
      tags:
      - Cart
...
swagger: "2.0"